requirements.txt
numpy>=1.24.0
pandas>=2.0.0
torch>=2.0.0
matplotlib>=3.7.0
PyYAML>=6.0.0
mlflow>=2.0.0
flask>=3.0.0
requests>=2.31.0
pyngrok>=7.0.0
